# doc/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: ""
  version: ""
paths:
  /api/sample/v1/sample_entities:
    get:
      description: List SampleEntity
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/DtoSampleEntityResponse'
                type: array
          description: OK
      summary: List SampleEntity
      tags:
      - SampleEntity
    post:
      description: Create SampleEntity
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DtoSampleEntityRequest'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DtoSampleEntityResponse'
          description: OK
      summary: Create SampleEntity
      tags:
      - SampleEntity
...
components:
  schemas:
    DtoSampleEntityRequest:
      properties:
        name:
          type: string
        text:
          type: string
      type: object
    DtoSampleEntityResponse:
      properties:
        id:
          minimum: 0
          type: integer
        name:
          type: string
        text:
          type: string
      type: object
...